api호출 백엔드로 옮김

+++ carbonlease/src/main/java/com/kh/openapi/sidebar/model/dto/StationAirResponse.java
@@ -6,15 +6,13 @@ import lombok.Data;
 @Data
 @Builder
 public class StationAirResponse {
-
-    private String stationName;   // 측정소명
-    private String sidoName;      // 시/도
-    private String dataTime;      // 측정일시
-
-    private Integer pm10;         // PM10 수치
-    private Integer pm25;         // PM2.5 수치
-    private Integer o3;           // 오존농도
-    private Integer co;           // 일산화탄소
-    private Integer khaiValue;    // 통합대기환경수치
-    private Integer khaiGrade;    // 통합대기환경지수 등급
+    private String stationName;
+    private String sidoName;
+    private String dataTime;
+    private int pm10;
+    private int pm25;
+    private double o3;
+    private double co;
+    private int khaiValue;
+    private int khaiGrade;
 }
